disk: add reconfigure_fstab().

+++ kvm/rhel/6/functions/distro.sh
@@ -422,8 +422,16 @@ function render_fstab() {
 	_EOS_
 }
 
+function reconfigure_fstab() {
+  local chroot_dir=$1
+  [[ -d "${chroot_dir}"    ]] || { echo "[ERROR] directory not found: ${chroot_dir} (distro:${LINENO})" >&2; return 1; }
+
+  render_fstab ${chroot_dir} | tee ${chroot_dir}/etc/fstab
+}
+
 function reconfigure_mtab() {
   local chroot_dir=$1
+  [[ -d "${chroot_dir}"    ]] || { echo "[ERROR] directory not found: ${chroot_dir} (distro:${LINENO})" >&2; return 1; }
 
   [[ -f ${chroot_dir}/etc/mtab ]] && rm -f ${chroot_dir}/etc/mtab || :
   run_in_target ${chroot_dir} ln -fs /proc/mounts /etc/mtab
